Removed tensorflow check before importing ROM_NN_Interface
